==> analysis/integration/merfish_rna/00.prepare.ipynb <==
import pandas as pd
import numpy as np
from wmb import *
from ALLCools.plot import *

import pathlib

rna_annot = aibs.get_smart_annot()

def save_index(cells, file_name, downsample=100000, random_state=0):
    cells = pd.Series(cells)
    if cells.size > downsample:
        cells = cells.sample(downsample, random_state=random_state)
    cells.to_csv(file_name, index=None, header=False)

## L1 - donwsample rna cells

#pathlib.Path('../L1/NonNeuron').mkdir(exist_ok=True)

save_index(rna_annot.get_index('cell'),
           file_name='./rna_cells.txt',
           downsample=100000,
           random_state=0)
